requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
asttokens==3.0.0
attrs==24.3.0
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
executing==2.1.0
Flask==2.3.3
frozenlist==1.5.0
gunicorn==21.2.0
icecream==2.1.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
multidict==6.1.0
packaging==24.2
propcache==0.2.1
Pygments==2.18.0
python-dotenv==1.0.0
PyYAML==6.0.2
requests==2.31.0
ua-parser==1.0.0
ua-parser-builtins==0.18.0.post1
urllib3==2.3.0
user-agents==2.2.0
Werkzeug==3.1.3
yarl==1.18.3
tabulate==0.9.0
newrelic==10.4.0
